Added some tests

+++ app/src/main/java/com/orange/devoxx/com/backend/beans/GroupModel.java
@@ -34,7 +34,7 @@ public class GroupModel {
         return mGroupId;
     }
 
-    public GroupModel setId(int id) {
+    public GroupModel setId(long id) {
         this.mGroupId = id;
         return this;
     }
